feat: Implement launch command for aiman package manager
This commit adds a new command called "launch", which allows the
user to execute installed AI projects easily. Currently, only
the "webui" project can be launched and all launch parameters
are hardcoded but this will be improved in future updates.

Future updates will include support for additional launch parameters
and projects, as well as the ability to customize launch settings.