feat: Pass command-line options to launched project
This commit adds functionality to the 'aiman' script that allows
command-line options and flags passed to the 'aiman launch [project]'
command to be forwarded to the launched project. This is an important
feature that enables users to customize the behavior of the launched
project according to their needs.

Please note that this feature is still in the testing phase and may
not work as expected.