feat: Automate creation of symbolic link 'webui/outputs' -> 'AIMan/Output'
This commit implements the automatic creation of a symbolic link
from 'webui/outputs' to 'AIMan/Output' as part of the webui project
installation.

The purpose of this feature is to centralize the storage location for
generated files (output) across projects. While currently implemented for
the 'webui' project, the intention is to extend this functionality to all
future projects administered by AIMan.