refactor: Refactor command parsing and project handling in 'aiman' script
This commit introduces a significant refactoring to the way the 'aiman'
interprets the user's desired command and the project on which the command
should be executed.

The new implementation allows users to run commands in the following format:
  ./aiman <project>.<command>.

For example, the command "./aiman webui.info" will now execute the "info"
command on the "webui" project. This refactoring is an important change,
and as a result, only the "info" command is currently functional. In
subsequent commits, the previously working commands will be reactivated.