feat: Implemented 'list' command and improved internal functions
This commit introduces the 'list' command in the aiman script,
allowing users to view the available AI projects they can install.
Additionally, several internal functions have been added to enhance
the development of new commands, as the new 'is_project_installed()'
function that simplifies the creation of new commands by providing
a reusable way to check if a project is already installed.