feat: Add 'ProjectPort' variable to force Gradio to listen on a specific port
This commit introduces the 'ProjectPort' variable that allows users
to specify a custom port for Gradio to listen on when running AI
projects with web interface.

Additionally, this commit includes minor text fixes and update to
the symlink for 'outputs' in AUTO1111 WebUI and Forges. These changes
improve the overall usability and reliability of the aiman script.