feat: Add support for managing multiple versions of the same project
This commit introduces the ability to install and manage multiple
versions of the same project. This is achieved by allowing the project
database to store multiple entries for the same project, each with a
different version identifier.

For example, a user can now install both "comfyui" and "comfyst:comfyui",
which will be managed by the same script (comfyui-handler.sh) but stored
in the database with different commit hashes and installation directories.
This allows users to easily switch between different versions of a project.