feat: add support for launching open-webui services in tmux
This commit adds the ability to launch the 3 services (ollama,
pipelines, and webui) from the 'aiman launch open-webui' command
inside a `tmux` session.

Previously, only screen was supported. Now, users can specify which
multiplexer to use with the `--screen` or `--tmux` options. If no option
is specified, it defaults to using `--tmux`, as it is considered a more
modern and maintained choice.